выполнил исправления
