--- overview/src/main/java/ru/pavlig43/overview/navigation/destination/CourseEdit.kt
@@ -1,17 +1,0 @@
-package ru.pavlig43.overview.navigation.destination
-
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.composable
-import kotlinx.serialization.Serializable
-import ru.pavlig.course_edit.ui.CourseEditingScreen
-
-
-
-fun NavGraphBuilder.courseEdit(
-    navController: NavHostController
-) {
-    composable<Destination.Edit> {
-        CourseEditingScreen()
-    }
-}
